CacheService + DataManager Fix
