+++ WeatherApp/WeatherApp/Services/WeatherRepository.swift
@@ -1,15 +1,15 @@
 //
-//  WeatherService.swift
+//  WeatherRepository.swift
 //  WeatherApp
 //
-//  Created by Phan Quyen on 13/08/2025.
+//  Created by Phan Quyen on 14/08/2025.
 //
 
 import Foundation
 import CoreLocation
 
-class WeatherService {
-    static let shared = WeatherService()
+class WeatherRepository {
+    static let shared = WeatherRepository()
     
     private let baseURL = "https://api.openweathermap.org/data/2.5"
     private let apiKey = "cdf3aca7aeb52390af4bc0db457f5f60"
@@ -105,8 +105,8 @@ class WeatherService {
     private func buildWeatherURL(latitude: Double, longitude: Double) -> URL? {
         var components = URLComponents(string: "\(baseURL)/weather")
         components?.queryItems = [
-            URLQueryItem(name: "lat", value: String(latitude)),
-            URLQueryItem(name: "lon", value: String(longitude)),
+            URLQueryItem(name: "lat", value: "\(latitude)"),
+            URLQueryItem(name: "lon", value: "\(longitude)"),
             URLQueryItem(name: "appid", value: apiKey),
             URLQueryItem(name: "units", value: "metric")
         ]
@@ -126,8 +126,8 @@ class WeatherService {
     private func buildForecastURL(latitude: Double, longitude: Double) -> URL? {
         var components = URLComponents(string: "\(baseURL)/forecast")
         components?.queryItems = [
-            URLQueryItem(name: "lat", value: String(latitude)),
-            URLQueryItem(name: "lon", value: String(longitude)),
+            URLQueryItem(name: "lat", value: "\(latitude)"),
+            URLQueryItem(name: "lon", value: "\(longitude)"),
             URLQueryItem(name: "appid", value: apiKey),
             URLQueryItem(name: "units", value: "metric")
         ]
@@ -144,34 +144,5 @@ class WeatherService {
         return grouped.map { _, dayItems in
             DailyForecast(from: dayItems)
         }.sorted { $0.date < $1.date }
-    }
-}
-
-enum WeatherAPIError: Error, LocalizedError {
-    case invalidURL
-    case noData
-    case networkError(NetworkError)
-    case decodingError(Error)
-    case invalidAPIKey
-    case cityNotFound
-    case rateLimitExceeded
-    
-    var errorDescription: String? {
-        switch self {
-        case .invalidURL:
-            return "Invalid URL"
-        case .noData:
-            return "No data received"
-        case .networkError(let networkError):
-            return "Network error: \(networkError.localizedDescription)"
-        case .decodingError(let error):
-            return "Decoding error: \(error.localizedDescription)"
-        case .invalidAPIKey:
-            return "Invalid API key"
-        case .cityNotFound:
-            return "City not found"
-        case .rateLimitExceeded:
-            return "API rate limit exceeded"
-        }
     }
 }
